refactor: enforce coding rules - repository layer, top-level imports, schema extraction
