+++ lambda/api/app/repositories/__init__.py
@@ -12,6 +12,8 @@ from .image_repository import (
     delete_images_by_app_name,
     delete_image,
     update_verification_status,
+    update_agent_status,
+    reset_processing_status,
     create_individual_page_record,
     update_parent_document_status,
     get_children_by_parent_id,
@@ -50,6 +52,8 @@ __all__ = [
     "delete_images_by_app_name",
     "delete_image",
     "update_verification_status",
+    "update_agent_status",
+    "reset_processing_status",
     "create_individual_page_record",
     "update_parent_document_status",
     "get_children_by_parent_id",
