+++ lambda/api/app/schemas/user.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+
+class UpdateProfileRequest(BaseModel):
+    display_name: str
